requirements.txt
Django==4.0.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-filter==21.1
pillow==9.3.0
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
pytest-django==4.5.2
django-cors-headers==3.11.0

python-dotenv==0.20.0
gunicorn==20.1.0

django-redis==5.2.0

pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.4

pytrends==4.9.2

google-cloud-bigquery==3.11.4

Pillow==10.0.0
yfinance==0.2.26
matplotlib==3.7.2

nltk==3.8.1
requests==2.31.0
newsapi-python==0.2.7
